added cli / tool for embodied-emissions
