--- cea/GUI/benchmark_graphs_tool.py
@@ -1,49 +1,0 @@
-import cea
-
-__author__ = "Daren Thomas"
-__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Daren Thomas"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-class BenchmarkGraphsTool(object):
-    def __init__(self):
-        self.label = 'Benchmark graphs'
-        self.description = 'Create benchmark plots of scenarios in a folder'
-        self.canRunInBackground = False
-
-    def getParameterInfo(self):
-        import arcpy
-        scenarios = arcpy.Parameter(
-            displayName="Path to the scenarios to plot",
-            name="scenarios",
-            datatype="DEFolder",
-            parameterType="Required",
-            direction="Input",
-            multiValue=True)
-        output_file = arcpy.Parameter(
-            displayName="Path to output PDF",
-            name="output_file",
-            datatype="DEFile",
-            parameterType="Required",
-            direction="Output")
-        output_file.filter.list = ['pdf']
-        return [scenarios, output_file]
-
-    def execute(self, parameters, messages):
-        import arcpy
-        scenarios = parameters[0].valueAsText
-        scenarios = scenarios.replace('"', '')
-        scenarios = scenarios.replace("'", '')
-        scenarios = scenarios.split(';')
-        arcpy.AddMessage(scenarios)
-        output_file = parameters[1].valueAsText
-
-        from cea.analysis import benchmark
-        reload(benchmark)
-        locator_list = [cea.inputlocator.InputLocator(scenario_path=scenario) for scenario in scenarios]
-        benchmark.benchmark(locator_list=locator_list, output_file=output_file)
-        return
