--- cea/GUI/emissions_tool.py
@@ -1,119 +1,0 @@
-import cea.inputlocator
-
-__author__ = "Daren Thomas"
-__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Daren Thomas"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-
-class EmissionsTool(object):
-    def __init__(self):
-        self.label = 'Emissions Operation'
-        self.description = 'Calculate emissions and primary energy due to building operation'
-        self.canRunInBackground = False
-
-    def getParameterInfo(self):
-        import arcpy
-        scenario_path = arcpy.Parameter(
-            displayName="Path to the scenario",
-            name="scenario_path",
-            datatype="DEFolder",
-            parameterType="Required",
-            direction="Input")
-        Qww_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to hot water consumption.",
-            name="Qww_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Qww_flag.value = True
-        Qhs_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to space heating.",
-            name="Qhs_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Qhs_flag.value = True
-        Qcs_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to space cooling.",
-            name="Qcs_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Qcs_flag.value = True
-        Qcdata_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to servers cooling.",
-            name="Qcdata_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Qcdata_flag.value = True
-        Qcrefri_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to refrigeration.",
-            name="Qcrefri_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Qcrefri_flag.value = True
-        Eal_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to appliances and lighting.",
-            name="Eal_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Eal_flag.value = True
-        Eaux_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to auxiliary electricity.",
-            name="Eaux_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Eaux_flag.value = True
-        Epro_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to electricity in industrial processes.",
-            name="Epro_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Epro_flag.value = True
-        Edata_flag = arcpy.Parameter(
-            displayName="Create a separate file with emissions due to electricity consumption in data centers.",
-            name="Edata_flag",
-            datatype="GPBoolean",
-            parameterType="Required",
-            direction="Input")
-        Edata_flag.value = True
-
-        return [scenario_path, Qww_flag, Qhs_flag, Qcs_flag, Qcdata_flag, Qcrefri_flag, Eal_flag, Eaux_flag, Epro_flag,
-                Edata_flag]
-
-    def isLicensed(self):
-        return True
-
-    def updateParameters(self, parameters):
-        return
-
-    def updateMessages(self, parameters):
-        return
-
-    def execute(self, parameters, messages):
-        from cea.analysis.operation import lca_operation
-        import cea.inputlocator
-        scenario_path = parameters[0].valueAsText
-        locator = cea.inputlocator.InputLocator(scenario_path)
-        Qww_flag = parameters[1].value
-        Qhs_flag = parameters[2].value
-        Qcs_flag = parameters[3].value
-        Qcdata_flag = parameters[4].value
-        Qcrefri_flag = parameters[5].value
-        Eal_flag = parameters[6].value
-        Eaux_flag = parameters[7].value
-        Epro_flag = parameters[8].value
-        Edata_flag = parameters[9].value
-        lca_operation(locator=locator, Qww_flag=Qww_flag, Qhs_flag=Qhs_flag, Qcs_flag=Qcs_flag, Qcdata_flag=Qcdata_flag,
-                      Qcrefri_flag=Qcrefri_flag, Eal_flag=Eal_flag, Eaux_flag=Eaux_flag, Epro_flag=Epro_flag,
-                      Edata_flag=Edata_flag)
